backend/pyproject.toml
[project]
name = "nova-backend"
version = "0.1.0"
description = "Nova Backend with PostgreSQL and LangChain tools"
authors = [
    {name = "Nova Team", email = "daniel.kuehlwein@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.2",
    "uuid6>=2023.5.2",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.1.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pydantic-settings>=2.9.1",
    "langchain-mcp-adapters>=0.1.4",
    "aiohttp>=3.12.8",
    "structlog>=24.1.0",
    "redis>=5.2.0",
    "orjson>=3.10.0",
    "pyyaml>=6.0.0",
    "watchdog>=4.0.0",
    "langchain-community>=0.3.25",
    "graphiti-core[google-genai]>=0.13.2",
    "zep-python>=2.0.2",
    "celery[redis]>=5.3.0",
    "flower>=2.0.1",
    "email-validator>=2.2.0",
]
requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "pytest-postgresql>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "database", "models", "tools", "agent"] 

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
]
